refactor: Refactor learner middleware and learner routes.
[skip ci]
