--- src/middleware/nemis/nhif.ts
@@ -1,3 +1,0 @@
-/*
- * Copyright (c) 2023. MIT License. Maina Derrick.
- */
